requirements.txt
python-telegram-bot==21.0
python-dotenv
web3
solana
schedule
